refactor(services): collapse _accumulated_helpers into accumulated
The helper file had one caller (accumulated.py itself) and a re-export
block that existed solely to preserve external import paths. Inlining
eliminates the indirection with no loss of depth.
