+++ src/quodeq/services/accumulated.py
@@ -3,32 +3,107 @@ from __future__ import annotations
 
 import threading
 from collections import OrderedDict
-from dataclasses import dataclass, field
+from dataclasses import dataclass, field, replace
 from pathlib import Path
 from typing import Any
 
-from quodeq.services.ports import RunInfo, list_runs
-from quodeq.core.types import DimensionResult
-from quodeq.shared.utils import _env_int
+from quodeq.core.scoring.internals import score_to_grade_label
+from quodeq.core.types import DimensionResult, to_camel_dict
 from quodeq.services._cache import make_lru_dimension_fetcher
-from quodeq.services.dim_resolution import is_eligible_for_default_view
 from quodeq.services.deleted import filter_deleted_from_dimensions
+from quodeq.services.dim_resolution import is_eligible_for_default_view
 from quodeq.services.dismissed import filter_dismissed_from_dimensions
 from quodeq.services._fs_projects import find_children as _find_children
+from quodeq.services.ports import RunInfo, calculate_trend, list_runs, most_frequent_grade, parse_numeric_score
+from quodeq.shared.utils import _env_int
 
-# Re-export helpers so existing external imports keep working.
+# Re-export so existing external imports keep working.
 from quodeq.services._accumulated_data import _read_all_run_data  # noqa: F401
-from quodeq.services._accumulated_helpers import (  # noqa: F401
-    _AccumulatedResult,
-    _aggregate_severity_counts,
-    _build_accumulated_response,
-    _compute_accumulated_scores,
-    _compute_accumulated_trends,
-    numeric_average,
-)
 
 _DEFAULT_ACC_CACHE_MAX = 256
 
+
+def numeric_average(dimensions: list[DimensionResult]) -> float | None:
+    """Compute the average numeric score from a list of DimensionResult objects."""
+    raw = [d.overall_score for d in dimensions if d.overall_score]
+    numeric = [s for s in (parse_numeric_score(v) for v in raw) if s is not None]
+    return round(sum(numeric) / len(numeric), 1) if numeric else None
+
+
+def _compute_accumulated_trends(
+    all_dimensions: list[DimensionResult],
+    prev_occurrence: dict[str, DimensionResult],
+) -> list[DimensionResult]:
+    result: list[DimensionResult] = []
+    for dim in all_dimensions:
+        dim_name = dim.dimension
+        previous = prev_occurrence.get(dim_name) if dim_name else None
+        trend = calculate_trend(
+            dim.overall_score,
+            previous.overall_score if previous else None,
+        )
+        result.append(replace(
+            dim,
+            trend=trend,
+            previous_run_id=previous.run_id if previous else None,
+            previous_score=previous.overall_score if previous else None,
+        ))
+    return result
+
+
+def _aggregate_severity_counts(all_dimensions: list[DimensionResult]) -> dict[str, int]:
+    """Sum violation/compliance counts and severity buckets across dimensions."""
+    total_violations = total_compliance = critical = major = minor = 0
+    for dim in all_dimensions:
+        totals = dim.totals
+        if totals:
+            total_violations += totals.violation_count
+            total_compliance += totals.compliance_count
+            critical += totals.severity.critical
+            major += totals.severity.major
+            minor += totals.severity.minor
+    return {
+        "totalViolations": total_violations, "totalCompliance": total_compliance,
+        "critical": critical, "major": major, "minor": minor,
+    }
+
+
+def _compute_accumulated_scores(
+    all_dimensions: list[DimensionResult], prev_run_latest: list[DimensionResult],
+) -> tuple[float | None, float | None]:
+    return numeric_average(all_dimensions), (numeric_average(prev_run_latest) if prev_run_latest else None)
+
+
+@dataclass(frozen=True)
+class _AccumulatedResult:
+    all_dimensions: list[DimensionResult]
+    dimensions_with_trend: list[DimensionResult]
+    severity: dict[str, int]
+    avg_score: float | None
+    prev_avg_score: float | None
+
+
+def _build_accumulated_response(project: str, result: _AccumulatedResult) -> dict[str, Any]:
+    return {
+        "project": project,
+        "dimensions": [to_camel_dict(d) for d in result.dimensions_with_trend],
+        "summary": {
+            "overallGrade": (
+                score_to_grade_label(result.avg_score) if result.avg_score is not None
+                else most_frequent_grade([d.overall_grade for d in result.all_dimensions if d.overall_grade])
+            ),
+            "numericAverage": result.avg_score,
+            "previousNumericAverage": result.prev_avg_score,
+            "totalViolations": result.severity["totalViolations"],
+            "totalCompliance": result.severity["totalCompliance"],
+            "dimensionCount": len(result.dimensions_with_trend),
+            "severity": {
+                "critical": result.severity["critical"],
+                "major": result.severity["major"],
+                "minor": result.severity["minor"],
+            },
+        },
+    }
 
 def create_accumulated_cache() -> tuple[OrderedDict[tuple, list[DimensionResult]], threading.Lock]:
     """Create the default accumulated-view LRU cache and its lock."""
